% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ac_mat}
\alias{get_ac_mat}
\title{get_ac_mat}
\usage{
get_ac_mat(n_var, max_cnt = 100)
}
\arguments{
\item{n_var}{total number of variants}

\item{max_cnt}{maximum allele count, default 100}
}
\value{
A `n_var`x`n_var` diagnal matrix of allele count information for `n_var` variants
}
\description{
simulation function: simulate allele count information for `n_var` variants, with a maximum allele count `max_cnt`
}
\examples{
ac_mat <- get_ac_mat(n_var=100, max_cnt = 100)
}
