% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_af_mat}
\alias{get_af_mat}
\title{get_af_mat}
\usage{
get_af_mat(AC, n_ind)
}
\arguments{
\item{AC}{a diagonal matrix of allele count information for all variants}

\item{n_ind}{total number of individuals in the population}
}
\value{
A `n_var`x`n_var` diagnal matrix of allele frequency information for `n_var` (dimension of `AC`) variants
}
\description{
simulation function: compute allele frequency information variants with allele counts stored in diagonal matrix `AC` from a population of sample size `n_ind`
}
\examples{
af_mat <- get_af_mat(AC = c(20, 50, 10, 1, 5), n_ind = 10000)
}
