% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geno_mat}
\alias{get_geno_mat}
\title{get_geno_mat}
\usage{
get_geno_mat(AC, n_ind)
}
\arguments{
\item{AC}{allele counts of loci (length `m`)}

\item{n_ind}{total number of indicitions}
}
\value{
An `n_ind`x`m` matrix of genotype information of `n_ind` individuals and `m` variants
}
\description{
simulation function: simulate genotype information for a set of loci with allele counts `AC`
}
\examples{
geno_mat <- get_geno_mat(AC = c(20, 50, 10, 1, 5), n_ind = 10000)
}
