% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_likelihood_test_stats}
\alias{get_likelihood_test_stats}
\title{get_likelihood_test_stats}
\usage{
get_likelihood_test_stats(n_ind, r, b1_hat, b2_hat, c, A)
}
\arguments{
\item{n_ind}{total number of individuals}

\item{r}{phenotypic correlation between the two phenotypes}

\item{b1_hat}{estimated effect size of the first phenotype across all variants}

\item{b2_hat}{estimated effect size of the second phenotype across all variants}

\item{c}{MLE of the slope between the two sets of variant effect sizes}

\item{A}{Allele frequency information}
}
\value{
A single numeric value representing the test statistic of ALLSPICE (maximum likelihood ratio)
}
\description{
ALLSPICE function: compute the maximum likelihood ratio of the ALLSPICE test statistic
}
