% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_single_geno}
\alias{get_single_geno}
\title{get_single_geno}
\usage{
get_single_geno(cnt, n_ind)
}
\arguments{
\item{cnt}{number of individuals with the mutation}

\item{n_ind}{total number of individuals}
}
\value{
A binary vector representing the genotype information of `n_ind` individuals for a particular locus, where `cnt` entries has value 1.
}
\description{
simulation function: simulate genotype information for one locus, where `cnt` samples out of `n_ind` has the mutation
}
\examples{
geno <- get_single_geno(cnt = 100, n_ind = 10000)
}
